requirements.txt
openai>=1.0.0
chromadb>=0.5.0
pyyaml>=6.0
azure-kusto-data>=4.3.0
azure-identity>=1.15.0
rich>=13.7.0
prompt-toolkit>=3.0.0
gitpython>=3.1.0
sentence-transformers>=3.0.0
fastapi>=0.115.0
uvicorn>=0.34.0
botbuilder-core>=4.16.0
aiohttp>=3.9.0
